// frontend/src/app/api/model/contact.ts
/**
 * API Agenda de Contactos
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Representación de un contacto.
 */
export interface Contact { 
    readonly id?: number;
    nombre: string;
    apellido: string;
    telefono?: string;
    correo: string;
    direccion?: string;
    empresa?: string;
    cargo?: string;
    categoria?: string;
    notas?: string;
    favorito?: boolean;
    readonly fechaCreacion?: string;
    readonly fechaActualizacion?: string;
}